Added Bounus

+++ LessonDetails.java
@@ -2,10 +2,10 @@ package emotionalSupport;
 
 public class LessonDetails {
     public int no_of_lecs, no_of_sections;
-    public float lec_hrs, sec_hrs;
+    public int lec_hrs, sec_hrs;
     public int lectureAttendees, sectionAttendees;
     
-	public LessonDetails(int no_of_lecs, int no_of_sections, float lec_hrs, float sec_hrs, int lectureAttendees,
+	public LessonDetails(int no_of_lecs, int no_of_sections, int lec_hrs, int sec_hrs, int lectureAttendees,
 			int sectionAttendees) 
 	{
 		this.no_of_lecs = no_of_lecs;
@@ -16,7 +16,7 @@ public class LessonDetails {
 		this.sectionAttendees = sectionAttendees;
 	}
 	
-	public LessonDetails(int no_of_lecs, int no_of_sections, float lec_hrs, float sec_hrs)
+	public LessonDetails(int no_of_lecs, int no_of_sections, int lec_hrs, int sec_hrs)
 	{
 		this(no_of_lecs, no_of_sections, lec_hrs, sec_hrs, 0, 0);
 	}
